+++ packages/wesjet/src/cli/commands/BuildCommand.ts
@@ -22,7 +22,9 @@ export class BuildCommand extends BaseCommand {
     pipe(
       this.clearCacheIfNeeded(),
       T.chain(() => core.getConfig({ configPath: this.configPath })),
-      T.tap((config) => (config.source.options.disableImportAliasWarning ? T.unit : T.fork(core.validateTsconfig))),
+      T.tap((config) =>
+        config.source.options.disableImportAliasWarning ? T.unit : T.fork(core.validateTsconfig),
+      ),
       T.chain((config) => core.generateDotpkg({ config, verbose: this.verbose })),
       T.tap(core.logGenerateInfo),
       OT.withSpan('@wesjet/cli/commands/BuildCommand:executeSafe'),