+++ packages/wesjet/src/client/index.ts
@@ -1,1 +1,2 @@
-export * from '@wesjet/client'
+export * from './guards.js'
+export * from './utils.js'