+++ packages/wesjet/src/client/utils.ts
@@ -7,7 +7,10 @@ type PickRequired<T> = {
 }
 type ConvertPick<T> = ConvertUndefined<T> & PickRequired<T>
 
-export const pick = <Obj, Keys extends keyof Obj>(obj: Obj, keys: Keys[]): ConvertPick<{ [K in Keys]: Obj[K] }> => {
+export const pick = <Obj, Keys extends keyof Obj>(
+  obj: Obj,
+  keys: Keys[],
+): ConvertPick<{ [K in Keys]: Obj[K] }> => {
   return keys.reduce((acc, key) => {
     acc[key] = obj[key]
     return acc