+++ packages/@wesjet/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @wesjet/utils
 
+## 0.0.19
+
+### Patch Changes
+
+- Wesjet
+
+## 0.0.18
+
+### Patch Changes
+
+- Automated
+
 ## 0.0.17
 
 ### Patch Changes