+++ packages/wesjet/CHANGELOG.md
@@ -1,5 +1,15 @@
 # wesjet
 
+## 0.0.19
+
+### Patch Changes
+
+- Wesjet
+- Updated dependencies
+  - @wesjet/maker@0.0.19
+  - @wesjet/utils@0.0.19
+  - @wesjet/core@0.0.19
+
 ## 0.0.17
 
 ### Patch Changes